backend/requirements.txt
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
SQLAlchemy==2.0.17
python-dotenv==0.21.0
pytest==7.4.0
psycopg2==2.9.6
beautifulsoup4==4.12.2
requests==2.31.0
gunicorn==21.2.0